savvyr: Survival Analysis for AdVerse Events with VarYing Follow-Up Times

The SAVVY (Survival Analysis for AdVerse Events with VarYing Follow-Up Times) project is a consortium of academic and pharmaceutical industry partners that aims to improve the analyses of adverse event (AE) data in clinical trials through the use of survival techniques appropriately dealing with varying follow-up times and competing events, see Stegherr, Schmoor, Beyersmann, et al. (2021) doi:10.1186/s13063-021-05354-x . Although statistical methodologies have advanced, in AE analyses often the incidence proportion, the incidence density or a non-parametric Kaplan-Meier estimator are used, which either ignore censoring or competing events. This package contains functions to easily conduct the proposed improved AE analyses.

See also

Author

Maintainer: Thomas Kuenzel thomas.kuenzel@roche.com

Authors:

Kaspar Rufibach kaspar.rufibach@roche.com

Regina Stegherr regina.stegherr@uni-ulm.de

Daniel Sabanés Bové daniel@rconis.com

Other contributors:

F. Hoffmann-La Roche AG [copyright holder, funder]

Universität Ulm [copyright holder, funder]